+++ webman-templates.php
@@ -3,13 +3,13 @@
  * Plugin Name:        WebMan Templates
  * Plugin URI:         https://www.webmandesign.eu/portfolio/webman-templates-wordpress-plugin/
  * Description:        Provides collection of pre-designed row templates for Beaver Builder page builder.
- * Version:            2.2.2
+ * Version:            2.2.3
  * Author:             WebMan Design, Oliver Juhas
  * Author URI:         https://www.webmandesign.eu/
  * License:            GNU General Public License v3
  * License URI:        http://www.gnu.org/licenses/gpl-3.0.txt
  * Requires at least:  4.5
- * Tested up to:       4.9
+ * Tested up to:       4.9.8
  *
  * @copyright  WebMan Design, Oliver Juhas
  * @license    GPL-3.0, https://www.gnu.org/licenses/gpl-3.0.html